_core/packages/python/openproblems/pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "openproblems"
dynamic = ["version"]
description = "OpenProblems core Python helper functions."
authors = [
  { name = "Robrecht Cannoodt", email = "rcannood@gmail.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">= 3.9, < 3.13"
dependencies = [
  'PyYAML'
]

[project.optional-dependencies]
test = [
  "pytest>=8.0"
]

[project.urls]
homepage = "https://openproblems.bio/documentation"
repository = "https://github.com/openproblems-bio/core"

[tool.setuptools.packages.find]
where = ["src"]
include = ["openproblems*"]
exclude = ["tests*"]

[tool.setuptools_scm]
root = "../../.."
